change to ts
