+++ index.js
@@ -1,4 +1,1 @@
-(function() {
-	this.version = "0.0.1";
-	this.utf8 = require('./lib/utf8');
-}).call(module.exports);
+module.exports = require('./lib').default;
