+++ lib/index.js
@@ -1,0 +1,6 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+var index_1 = require("./utf8/index");
+exports.default = {
+    utf8: index_1.default
+};
