+++ lib/utf8/index.js
@@ -1,0 +1,156 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+var Utf8;
+(function (Utf8) {
+    var stringFromCharCode = String.fromCharCode;
+    function ucs2encode(array) {
+        var length = array.length;
+        var index = -1;
+        var value;
+        var output = '';
+        while (++index < length) {
+            value = array[index];
+            if (value > 0xFFFF) {
+                value -= 0x10000;
+                output += stringFromCharCode(value >>> 10 & 0x3FF | 0xD800);
+                value = 0xDC00 | value & 0x3FF;
+            }
+            output += stringFromCharCode(value);
+        }
+        return output;
+    }
+    Utf8.ucs2encode = ucs2encode;
+    ;
+    function ucs2decode(str) {
+        var output = [];
+        var counter = 0;
+        var length = str.length;
+        var value;
+        var extra;
+        while (counter < length) {
+            value = str.charCodeAt(counter++);
+            if (value >= 0xD800 && value <= 0xDBFF && counter < length) {
+                extra = str.charCodeAt(counter++);
+                if ((extra & 0xFC00) == 0xDC00) {
+                    output.push(((value & 0x3FF) << 10) + (extra & 0x3FF) + 0x10000);
+                }
+                else {
+                    output.push(value);
+                    counter--;
+                }
+            }
+            else {
+                output.push(value);
+            }
+        }
+        return output;
+    }
+    Utf8.ucs2decode = ucs2decode;
+    function createByte(codePoint, shift) {
+        return stringFromCharCode(((codePoint >> shift) & 0x3F) | 0x80);
+    }
+    function encodeCodePoint(codePoint) {
+        if ((codePoint & 0xFFFFFF80) == 0) {
+            return stringFromCharCode(codePoint);
+        }
+        var symbol = '';
+        if ((codePoint & 0xFFFFF800) == 0) {
+            symbol = stringFromCharCode(((codePoint >> 6) & 0x1F) | 0xC0);
+        }
+        else if ((codePoint & 0xFFFF0000) == 0) {
+            symbol = stringFromCharCode(((codePoint >> 12) & 0x0F) | 0xE0);
+            symbol += createByte(codePoint, 6);
+        }
+        else if ((codePoint & 0xFFE00000) == 0) {
+            symbol = stringFromCharCode(((codePoint >> 18) & 0x07) | 0xF0);
+            symbol += createByte(codePoint, 12);
+            symbol += createByte(codePoint, 6);
+        }
+        symbol += stringFromCharCode((codePoint & 0x3F) | 0x80);
+        return symbol;
+    }
+    function encode(str) {
+        var codePoints = ucs2decode(str);
+        var byteString = '';
+        for (var i = 0, len = codePoints.length; i < len; i++) {
+            var codePoint = codePoints[i];
+            byteString += encodeCodePoint(codePoint);
+        }
+        return byteString;
+    }
+    Utf8.encode = encode;
+    function decode(byteString) {
+        var byteArray = ucs2decode(byteString);
+        var byteCount = byteArray.length;
+        var byteIndex = 0;
+        var codePoints = [];
+        var tmp;
+        function readContinuationByte() {
+            if (byteIndex >= byteCount) {
+                throw Error('Invalid byte index');
+            }
+            var continuationByte = byteArray[byteIndex] & 0xFF;
+            byteIndex++;
+            if ((continuationByte & 0xC0) == 0x80) {
+                return continuationByte & 0x3F;
+            }
+            throw Error('Invalid continuation byte');
+        }
+        function decodeSymbol() {
+            var byte1;
+            var byte2;
+            var byte3;
+            var byte4;
+            var codePoint;
+            if (byteIndex > byteCount) {
+                throw Error('Invalid byte index');
+            }
+            if (byteIndex == byteCount) {
+                return false;
+            }
+            byte1 = byteArray[byteIndex] & 0xFF;
+            byteIndex++;
+            if ((byte1 & 0x80) == 0) {
+                return byte1;
+            }
+            if ((byte1 & 0xE0) == 0xC0) {
+                var byte2_1 = readContinuationByte();
+                codePoint = ((byte1 & 0x1F) << 6) | byte2_1;
+                if (codePoint >= 0x80) {
+                    return codePoint;
+                }
+                else {
+                    throw Error('Invalid continuation byte');
+                }
+            }
+            if ((byte1 & 0xF0) == 0xE0) {
+                byte2 = readContinuationByte();
+                byte3 = readContinuationByte();
+                codePoint = ((byte1 & 0x0F) << 12) | (byte2 << 6) | byte3;
+                if (codePoint >= 0x0800) {
+                    return codePoint;
+                }
+                else {
+                    throw Error('Invalid continuation byte');
+                }
+            }
+            if ((byte1 & 0xF8) == 0xF0) {
+                byte2 = readContinuationByte();
+                byte3 = readContinuationByte();
+                byte4 = readContinuationByte();
+                codePoint = ((byte1 & 0x0F) << 0x12) | (byte2 << 0x0C) |
+                    (byte3 << 0x06) | byte4;
+                if (codePoint >= 0x010000 && codePoint <= 0x10FFFF) {
+                    return codePoint;
+                }
+            }
+            throw Error('Invalid UTF-8 detected');
+        }
+        while ((tmp = decodeSymbol()) !== false) {
+            codePoints.push(tmp);
+        }
+        return ucs2encode(codePoints);
+    }
+    Utf8.decode = decode;
+})(Utf8 || (Utf8 = {}));
+exports.default = Utf8;
