+++ package.json
@@ -1,16 +1,19 @@
 {
-	"name":"latte_encode",
-	"version":"0.0.1",
-	"description":"latte_encode",
-	"keywords":["latte"],
-	"author":"qiongtubao",
-	"main":"./index.js",
-	"repository":{
-		"type":"git",
-		"url":"https://github.com/qiongtubao/latte_encode.git"
+	"name": "latte_encode",
+	"version": "0.0.1",
+	"description": "latte_encode",
+	"keywords": [
+		"latte",
+		"latte_encode"
+	],
+	"author": "qiongtubao",
+	"main": "lib/index.js",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/qiongtubao/latte_encode.git"
 	},
-	"dependencies":{
-		
-	}
-	
+	"dependencies": {
+		"latte_lib": "*"
+	},
+	"devDependencies": {}
 }
