+++ src/index.ts
@@ -1,0 +1,5 @@
+import utf8 from "./utf8/index"
+
+export default {
+  utf8
+}
