+++ test/utf8/index.js
