+++ tsconfig.json
@@ -1,0 +1,16 @@
+{
+      "compilerOptions": {
+            "target": "es5",
+            "noImplicitAny": false,
+            "module": "commonjs",
+            "removeComments": true,
+            "sourceMap": false,
+            "outDir": "lib"
+      },
+     "include":[
+         "src"
+     ],
+     "exclude": [
+         
+     ]
+}
